+++ CHANGES.md
@@ -7,6 +7,7 @@
 - Upgrade logback to 1.2.10
 - #3294 Indexer will auto-fixes non-topologically closed Polygons / MultiPolygons
 - Crawler will now wait for 3 seconds (used to be 1 second) before perform trim action to avoid timeout issue
+- Refactor Indexer code to remove the redundant index queue
 
 ## 1.1.0